Renamed configuration_managers package to configurationsmanagers
Apparently Golang linting is not fond of the underscores in package names.
