+++ internal/app/secretless/configurationmanagers/kubernetes/crd/constants.go
